fix compilation on mac
